Air Cargo Analysis
You, as a DBA expert, need to focus on identifying the regular customers to provide
offers, analyze the busiest route which helps to increase the number of aircraft
required and prepare an analysis to determine the ticket sales details. This will
ensure that the company improves its operability and becomes more customer-
centric and a favorable choice for air travel.